requirements_dev.txt
-r requirements.txt

django_debug_toolbar==1.8
django-nose==1.4.4
nose==1.3.7
rednose==1.2.2

ipython==6.1.0

